+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
+## 1.0.0 (X July 2022)
+
+### Added
+- Added the creation_plan block in the subscription resource schema.
+
+### Removed
+- Removed the database block from the subscription resource schema.
+
 ## 0.3.0 (May 24 2022)
 
 ### Added